added 3 options button and handleClick
