+++ src/components/dateIdea.js
@@ -1,0 +1,7 @@
+class DateIdea extends React.Component {
+  render() {
+    return (
+      <li>{this.props.value}</li>
+    );
+  }
+}
